Now avoiding to detect when only a small part of the color is in the image and fixing an error in partColored

+++ VisualAnalysis/ImageAnalyzer.cpp
@@ -66,5 +66,5 @@ double ImageAnalyzer::getElevation() const{
 }
 
 double ImageAnalyzer::partColored() const{
-  return nbPixels / (imageWidth * imageHeight);
+  return nbPixels / (double)(imageWidth * imageHeight);
 }
